// angular-frontend/src/app/clients/open-f1/models/session/session.interface.ts
export interface ISession {
    circuitKey: number
    circuitShortName: string
    countryCode: string
    countryKey: number
    countryName: string
    dateEnd: Date
    dateStart: Date
    gmtOffset: number
    location: string
    meetingKey: number
    sessionKey: number
    sessionName: string
    sessionType: string
    year: number
}